requirements.txt
pillow
pyPS4Controller